requirements.txt
arxiv
sqlalchemy
requests
eralchemy
openai
python-dotenv
graphviz
numpy
scikit-learn